Move the Server class to stlite_server.server
